+++ NEWS.md
@@ -2,6 +2,7 @@
 **+**: Denotes potentially code-breaking changes
 
 ## Changes in functionality
+o `distance()` now works for calculation of distances between two `GVector`s or a `GVector` and itself.
 o **+** `extract()` and **+** `spatSample()`: Changed default value of `cats` argument to `TRUE`.
 
 # fasterRaster 8.3.0.7019 (2024-06-08)