+++ NEWS.md
@@ -1,6 +1,9 @@
 # fasterRaster 8.3.0.7020 (2024-06-XX)
 **+**: Denotes potentially code-breaking changes
 
+## New functions and functionality
+o `sineRast()`: Creates sine wave rasters.
+
 ## Changes in functionality
 o `distance()` now works for calculation of distances between two `GVector`s or a `GVector` and itself.
 o **+** `extract()` and **+** `spatSample()`: Changed default value of `cats` argument to `TRUE`.